refactor state store
