+++ src/knowlang/indexing/state_store/postgres.py
@@ -1,10 +1,3 @@
-from __future__ import annotations
-
-import hashlib
-from datetime import datetime
-from pathlib import Path
-from typing import Dict, List, Optional, Set
-
 from knowlang.configs.state_store_config import StateStoreConfig
 from knowlang.core.types import StateStoreProvider
 from knowlang.indexing.state_store.base import StateChangeType, StateStore
